backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pillow==11.0.0
pillow-heif==0.18.0
python-multipart==0.0.17